+++ src/openai/_utils/_typing.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import Any, TypeVar, cast
+from typing import Any, TypeVar, Iterable, cast
+from collections import abc as _c_abc
 from typing_extensions import Required, Annotated, get_args, get_origin
 
 from .._types import InheritsGeneric
@@ -13,6 +14,12 @@ def is_annotated_type(typ: type) -> bool:
 
 def is_list_type(typ: type) -> bool:
     return (get_origin(typ) or typ) == list
+
+
+def is_iterable_type(typ: type) -> bool:
+    """If the given type is `typing.Iterable[T]`"""
+    origin = get_origin(typ) or typ
+    return origin == Iterable or origin == _c_abc.Iterable
 
 
 def is_union_type(typ: type) -> bool:
@@ -45,7 +52,13 @@ def extract_type_arg(typ: type, index: int) -> type:
         raise RuntimeError(f"Expected type {typ} to have a type argument at index {index} but it did not") from err
 
 
-def extract_type_var_from_base(typ: type, *, generic_bases: tuple[type, ...], index: int) -> type:
+def extract_type_var_from_base(
+    typ: type,
+    *,
+    generic_bases: tuple[type, ...],
+    index: int,
+    failure_message: str | None = None,
+) -> type:
     """Given a type like `Foo[T]`, returns the generic type variable `T`.
 
     This also handles the case where a concrete subclass is given, e.g.
@@ -104,4 +117,4 @@ def extract_type_var_from_base(typ: type, *, generic_bases: tuple[type, ...], in
 
         return extracted
 
-    raise RuntimeError(f"Could not resolve inner type variable at index {index} for {typ}")
+    raise RuntimeError(failure_message or f"Could not resolve inner type variable at index {index} for {typ}")