+++ src/openai/resources/files.py
@@ -238,6 +238,7 @@ class Files(SyncAPIResource):
         """
         if not file_id:
             raise ValueError(f"Expected a non-empty value for `file_id` but received {file_id!r}")
+        extra_headers = {"Accept": "application/binary", **(extra_headers or {})}
         return self._get(
             f"/files/{file_id}/content",
             options=make_request_options(
@@ -272,7 +273,6 @@ class Files(SyncAPIResource):
         """
         if not file_id:
             raise ValueError(f"Expected a non-empty value for `file_id` but received {file_id!r}")
-        extra_headers = {"Accept": "application/json", **(extra_headers or {})}
         return self._get(
             f"/files/{file_id}/content",
             options=make_request_options(
@@ -511,6 +511,7 @@ class AsyncFiles(AsyncAPIResource):
         """
         if not file_id:
             raise ValueError(f"Expected a non-empty value for `file_id` but received {file_id!r}")
+        extra_headers = {"Accept": "application/binary", **(extra_headers or {})}
         return await self._get(
             f"/files/{file_id}/content",
             options=make_request_options(
@@ -545,7 +546,6 @@ class AsyncFiles(AsyncAPIResource):
         """
         if not file_id:
             raise ValueError(f"Expected a non-empty value for `file_id` but received {file_id!r}")
-        extra_headers = {"Accept": "application/json", **(extra_headers or {})}
         return await self._get(
             f"/files/{file_id}/content",
             options=make_request_options(