+++ src/openai/types/audio/transcription_create_params.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Union
+from typing import List, Union
 from typing_extensions import Literal, Required, TypedDict
 
 from ..._types import FileTypes
@@ -49,4 +49,12 @@ class TranscriptionCreateParams(TypedDict, total=False):
     0.2 will make it more focused and deterministic. If set to 0, the model will use
     [log probability](https://en.wikipedia.org/wiki/Log_probability) to
     automatically increase the temperature until certain thresholds are hit.
+    """
+
+    timestamp_granularities: List[Literal["word", "segment"]]
+    """The timestamp granularities to populate for this transcription.
+
+    Any of these options: `word`, or `segment`. Note: There is no additional latency
+    for segment timestamps, but generating word timestamps incurs additional
+    latency.
     """